requirements.txt
pandas==2.0.2
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
albumentations==0.4.3
opencv-python==4.6.0.66
pudb==2019.2
imageio==2.9.0
imageio-ffmpeg==0.4.2
pytorch-lightning==2.1.2
torchmetrics==1.2.1
omegaconf==2.1.1
test-tube>=0.7.5
streamlit>=0.73.1
einops==0.3.0
transformers==4.19.2
webdataset==0.2.5
open-clip-torch==2.7.0
gradio==3.13.2
kornia==0.6
invisible-watermark>=0.1.5
streamlit-drawable-canvas==0.8.0
timm==0.5.4
xformers==0.0.20
clip-anytorch==2.5.2
scikit-learn==1.2.2
python-dotenv==1.0.0
mne==1.4.2
pytest==7.4.0
PyWavelets==1.4.1
SQLAlchemy==2.0.20
alembic==1.12.0
pyodbc==4.0.39
pysqlite3==0.5.2
tensorboard==2.15.1
braindecode==0.8.1
torch_geometric==2.5.0
seaborn==0.13.2
